feat: 임시 SideBar 완성
