--- src/components/DashBoard/SideMenu/DashboardList/DashboardList.module.scss
@@ -1,1 +1,0 @@
-@import '@/src/styles/global.scss';
