--- src/components/DashBoard/SideMenu/DashboardList/index.tsx
@@ -1,11 +1,0 @@
-import ListHeader from './ListHeader/ListHeader'
-
-const DashboardList = () => {
-  return (
-    <div>
-      <ListHeader />
-    </div>
-  )
-}
-
-export default DashboardList
