+++ src/components/DashBoard/SideMenu/DashboardListSection/DashboardList/DashboardList.module.scss
@@ -1,0 +1,34 @@
+@import '@/src/styles/global.scss';
+
+.wrapper {
+  display: flex;
+  align-items: center;
+  padding: 1.2rem 0 1.2rem 1.2rem;
+
+  @include mobile {
+    padding: 2rem 0;
+    justify-content: center;
+  }
+}
+
+.dashboard-color {
+  width: 0.8rem;
+  height: 0.8rem;
+  border-radius: 100px;
+  background-color: $color-green;
+}
+
+.dashboard-title {
+  margin: 0 0.6rem 0 1.6rem;
+  @include font-style(1.8rem, 500, normal);
+
+  @include mobile {
+    display: none;
+  }
+}
+
+.crown {
+  @include mobile {
+    display: none;
+  }
+}
