+++ src/components/DashBoard/SideMenu/DashboardListSection/DashboardListSection.module.scss
@@ -1,0 +1,7 @@
+@import '@/src/styles/global.scss';
+
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
