+++ src/components/DashBoard/SideMenu/constants.js
@@ -3,3 +3,5 @@ export const LOGO_IMAGE = '/icons/logo.svg'
 export const LOGO_TEXT = '/icons/taskify.svg'
 
 export const ADD_BOX = '/icons/add-box.svg'
+
+export const CROWN = '/icons/crown-icon.svg'
